fix(web): refactor amount handling in invoice creation and payout columns; ensure proper parsing of amounts and enhance token display logic

+++ apps/web/src/app/api/invoice/[user]/[invoice]/route.ts
@@ -79,7 +79,7 @@ export async function PUT(
 
         const currentBalance = currentUser?.totalBalance || "0";
         const newBalance = (
-          parseInt(currentBalance) + parseInt(payload.amountPaidInCents)
+          parseFloat(currentBalance) + parseFloat(payload.amountPaidInCents)
         ).toString();
 
         await tx.user.update({
